Add mempool config.rs file
